Moving mock web server logic to HttpBaseTestCase in order to localize and fix the testing of API calls
